# Refactor of graduation job test