… spec/jobs/graduation_job_spec.rb