* Do not set expectations in a before block
* Put all expectations in one test so we don't run the before block over and over again
* Document exactly what we expect to happen when we run the graduation job
* Combine the test for workflow transition with the other tests
* Improved comments and clarity for what we're testing for
* Improved performance

```
Before:
Finished in 2 minutes 52.2 seconds (files took 10.2 seconds to load)
After:
Finished in 1 minute 9.3 seconds (files took 10.21 seconds to load)
```